Refactor: core-color-input -> color-input
