+++ docs/docs/palettes/app/vue-components/color-input.js
