+++ docs/docs/palettes/custom.njk
@@ -11,7 +11,7 @@ description: Create your own color palette from scratch, from one or more seed c
 
 <div id="seed-colors">
 	<template v-for="color, i in seedColors">
-		<core-color-input v-model="seedColors[i]" @delete="seedColors.splice(i, 1)"></core-color-input>
+		<color-input v-model="seedColors[i]" @delete="seedColors.splice(i, 1)"></color-input>
 	</template>
 	<wa-button class="add-button" appearance="outlined" @click="seedColors.push(seedColorSamples.shift() || '')">
 		<wa-icon slot="prefix" name="plus" variant="regular"></wa-icon>
